project 3

+++ 03.Project.py
@@ -1,6 +1,6 @@
-x = int(input("Enter first number: "))
+x = float(input("Enter first number: "))
 o = str(input("Enter operator (+,-,*,/): "))
-y = int(input("Enter second number: "))
+y = float(input("Enter second number: "))
 if o == "+":
     print(str(x) + " + " + str(y) + " = " + (str(x + y)))
 elif o == "-":
